Fix lux sensor readings being incorrect
I had the LSB and the MSB swapped when calculating the floating point value of the lux

+++ lib/lux_sensor/lux_sensor.h
@@ -11,6 +11,6 @@
 #include "repo/lib/drivers/i2c_driver.h"
 
 void lux_sensor_init(uint8_t I2Cx);
-uint16_t lux_sensor_read_lux();
+float lux_sensor_read_lux();
 
 #endif /* REPO_LIB_LUX_SENSOR_LUX_SENSOR_H_ */
